Fix clippy warnings without semantic differences
This commit contains fixed clippy warnings that should not modify the
behavior of the code.

+++ talpid-core/src/routing/mod.rs
@@ -86,7 +86,7 @@ impl RequiredRoute {
             node: node.into(),
             prefix,
             #[cfg(target_os = "linux")]
-            table_id: u32::from(crate::linux::TUNNEL_TABLE_ID),
+            table_id: crate::linux::TUNNEL_TABLE_ID,
         }
     }
